fix: refacto and bugfix ingestor
fix: add some configs to ingestor

fix: remove logs

Revert "fix: remove batch"

This reverts commit b497ce7f6581c73fc51312dbfda4385a3fcc710c.

fix: remove batch

fix: test log
