+++ .gitignore
@@ -1,7 +1,6 @@
 /target
 .idea
 .env
-db/
 infra/db/password.txt
 
 # Vscode settings
