+++ pragma-ingestor/src/db/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod insert;
+pub(crate) mod process;
